requirements.txt
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.7.0
python-dotenv>=0.19.0
pyyaml>=6.0
